+++ src/tsconfig.browser.json
@@ -1,21 +1,5 @@
 {
-  "compilerOptions": {
-    "module": "es2015",
-    "moduleResolution": "node",
-    "target": "es5",
-    "noImplicitAny": false,
-    "sourceMap": false,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "lib": [
-      "es2015",
-      "dom"
-    ],
-    "outDir": "lib",
-    "typeRoots": [
-      "../node_modules/@types"
-    ]
-  },
+  "extends": "../tsconfig.json",
   "exclude": [
     "./main.server.ts",
     "./main.server.aot.ts",
