--- webpack/helpers.js
@@ -1,9 +1,0 @@
-const { resolve } = require('path');
-
-function root(path) {
-  return resolve(__dirname, '..', path);
-}
-
-module.exports = {
-  root: root
-};
