--- webpack/webpack.server.js
@@ -1,10 +1,0 @@
-const { root } = require('./helpers');
-const { AotPlugin } = require('@ngtools/webpack');
-
-module.exports = {
-  entry: root('./src/main.server.ts'),
-  output: {
-    filename: 'server.js'
-  },
-  target: 'node'
-};
